src/requirements.txt
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
